fixed prompthandler

+++ Doctor Game/Assets/Scripts/PromptHandler.cs
@@ -29,7 +29,7 @@ public class PromptHandler : MonoBehaviour
         prompt.transform.GetChild(4).GetComponent<Text>().text = "Time: " + timeCost;
         prompt.transform.GetChild(7).GetComponent<Text>().text = "Stress: " + stressBuff;
 
-        data = new PromptData(float.Parse(staminaCost.Substring(1)), int.Parse(stressCost.Substring(1)), int.Parse(timeCost.Substring(1, timeCost.IndexOf('h') - 1)));
+        data = new PromptData(float.Parse(staminaCost.Substring(1)), int.Parse(stressCost.Substring(1)), float.Parse(timeCost.Substring(1, timeCost.IndexOf('h') - 1)));
 
         prompt.gameObject.SetActive(true);
         targetScene = scene;
@@ -46,7 +46,7 @@ public class PromptHandler : MonoBehaviour
         Stats.Stress += data.stressCost;
         Debug.Log(data.timeCost);
         Debug.Log(data.timeCost*60);
-        Stats.UpdateTime(data.timeCost * 60, false);
+        Stats.UpdateTime((int)(data.timeCost * 60), false);
         if (targetScene != "")
         {
             SceneManager.LoadScene(targetScene);
@@ -60,7 +60,7 @@ struct PromptData
 {
     public int stressCost;
     public float staminaCost;
-    public int timeCost;
+    public float timeCost;
 
     public PromptData(float staC, int strC, int tC)
     {
